Adding login and profile views
I added two modals, one for login/signup, and one for updating status.

I also added styles and scripts for the little notification dot for availability.

Feel free to comment or uncomment:
`$('#signin-modal').modal('show');
` `$('#account-modal').modal('show');` to see the other modal.

My idea is the status will be a dropdown and the plus icon will prompt them to add another status. Maybe it'll be a modal with a checkbox for adding it to the status list or just as a one time. I don't know how the UI will work for deleting a status from the list, or deleting their account. We also have to figure out how the emoji UI would work. Ideally, something like facebook messenger's emoji picker for a group chat, where it has a long list of emojis and it's simple to pick one and update it.

Obviously colors need work, just going with a quick mockup right now so we can eventually get all 3 (or 4) views and then brainstorm what the user experience will be like.