Added homepage mockup
Obviously it's a bit ugly right now, just wanted to get my ideas on paper (well... on github).

Top right area shows your account info, click it to open your profile modal.

The main area shows all of the groups, with the expanded group showing the list of members, their statuses, and when theyw ere last updated. The non-expanded groups would just show basic quick info about the group. I think the expanded group should actually take up the majority of the screen space in the finished app. I was thinking the groups would act as accordions so one or zero open groups at a time, just click the name button to toggle them. We'll def have to make that obvious for the user.

The bottom seciton is a sticky footer with search, create group, and join group, all of which I was thinking would be modals.

I know this is a lot of thrown together sample code right now, but once we implement the scripting I think our code will look a lot cleaner. I'm mostly just worried about the backend for this project, but my current mockup is ugly enough / has enough room for improvement for Alex/Nisha you two to have some stuff to work on while Sean and I set up backend.